move selenium tests to journey folder
