+++ tests/System/journeys/example.js
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-const {AbstractTest} = require("./testFramework");
+const {AbstractTest} = require("../testFramework");
 
 class Example extends AbstractTest {
